requirements.txt
slack-bolt
anthropic
flask
python-dotenv
